[#108352462] Developers can easily seed data (#1)
This is the first part of a change to allow
developers to more easily seed survey data. The
change reworks the real world sharing survey seed
data into one large file.  The code now creates
survey items such as facebook questions, and then
waits until the database has saved the question
before inserting a new set of records that
depend on that question.