Fixed Server Tests, Refreshing Angular
On a page refresh, the front end would not
reload properly.  Adding a route to handle
anything that is not the api.  The * path
will return the index.html of the angular
app.

Also, the tests work again.  Paths needed updating.

Finally, the README was updated to reflect the
setup changes.