Adding 500 response failed get survey
The get request for a survey may fail
if the seeded data is bad.  Sending
back a 500 so the reason for the error
is more obvious.

One small addition to the commit is to
redirect the angular code to the homepage
once a survey is submitted so the user knows
something has happened.